Start integration of Route into MapViewState
Route struct was moved into its own file. Additionally the RouteTable was moved into MapViewState and renamed mRouteTable.
