+++ OPHD/States/MapViewState.h
@@ -5,6 +5,8 @@
 #include "Wrapper.h"
 
 #include "Planet.h"
+#include "Route.h"
+
 #include "../Common.h"
 #include "../Constants.h"
 
@@ -253,9 +255,11 @@ private:
 	ReportsUiCallback mReportsUiCallback;
 	MapChangedCallback mMapChangedCallback;
 
-	// MISCELLANEOUS
+	// ROUTING
 	micropather::MicroPather* mPathSolver = nullptr;
+	std::map<class MineFacility*, Route> mRouteTable;
 
+	// MISCELLANEOUS
 	int mTurnCount = 0;
 
 	int mCurrentMorale = constants::DEFAULT_STARTING_MORALE;
